+++ core/src/sync_client.rs
@@ -15,6 +15,8 @@
 //!
 //! In case RPC is disabled, RPC calls will be skipped.
 
+#[cfg(feature = "multiproof")]
+use crate::types::MULTI_PROOF_CELL_DIMS;
 use crate::{
 	data::{
 		AchievedSyncConfidenceKey, BlockHeaderKey, Database, IsSyncedKey, LatestSyncKey,
@@ -29,6 +31,8 @@ use crate::{
 };
 
 use async_trait::async_trait;
+#[cfg(feature = "multiproof")]
+use avail_rust::utils::generate_multiproof_grid_dims;
 use avail_rust::{
 	kate_recovery::{commitments, matrix::Dimensions},
 	AvailHeader, H256,
@@ -154,7 +158,37 @@ async fn process_block(
 
 			// now this is in `u64`
 			let cell_count = rpc::cell_count_for_confidence(cfg.confidence);
-			let positions = rpc::generate_random_cells(dimensions, cell_count);
+			let positions = {
+				#[cfg(feature = "multiproof")]
+				{
+					let Some(multiproof_cell_dims) =
+						Dimensions::new(MULTI_PROOF_CELL_DIMS.0, MULTI_PROOF_CELL_DIMS.1)
+					else {
+						info!(
+							block_number,
+							"Skipping block with invalid multiproof cell dimensions",
+						);
+						return Ok(());
+					};
+
+					let Some(target_multiproof_grid_dims) =
+						generate_multiproof_grid_dims(multiproof_cell_dims, dimensions)
+					else {
+						info!(
+							block_number,
+							"Skipping block with invalid target multiproof grid dimensions",
+						);
+						return Ok(());
+					};
+
+					rpc::generate_random_cells(target_multiproof_grid_dims, cell_count)
+				}
+
+				#[cfg(not(feature = "multiproof"))]
+				{
+					rpc::generate_random_cells(dimensions, cell_count)
+				}
+			};
 
 			let (fetched, unfetched, _fetch_stats) = network_client
 				.fetch_verified(
@@ -267,7 +301,10 @@ mod tests {
 			header::extension::{v3::HeaderExtension, HeaderExtension::V3},
 			kate_commitment::v3::KateCommitment,
 		},
-		kate_recovery::{data::Cell, matrix::Position},
+		kate_recovery::{
+			data::{Cell, SingleCell},
+			matrix::Position,
+		},
 		subxt::config::substrate::Digest,
 		AvailHeader,
 	};
@@ -329,7 +366,7 @@ mod tests {
 			.returning(move |_, _, _, _, positions| {
 				let unfetched = vec![];
 				let fetched: Vec<Cell> = vec![
-					Cell {
+					Cell::SingleCell(SingleCell {
 						position: Position { row: 0, col: 0 },
 						content: [
 							183, 56, 112, 134, 157, 186, 15, 255, 245, 173, 188, 37, 165, 224, 226,
@@ -338,8 +375,8 @@ mod tests {
 							37, 200, 236, 4, 44, 40, 4, 3, 0, 11, 35, 249, 222, 81, 135, 1, 128, 0,
 							0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
 						],
-					},
-					Cell {
+					}),
+					Cell::SingleCell(SingleCell {
 						position: Position { row: 0, col: 2 },
 						content: [
 							153, 31, 34, 70, 221, 239, 97, 236, 3, 172, 44, 167, 114, 117, 186,
@@ -349,8 +386,8 @@ mod tests {
 							54, 133, 203, 162, 73, 252, 32, 42, 175, 24, 166, 142, 72, 226, 150,
 							163, 206, 115, 0,
 						],
-					},
-					Cell {
+					}),
+					Cell::SingleCell(SingleCell {
 						position: Position { row: 1, col: 1 },
 						content: [
 							146, 211, 61, 65, 166, 68, 252, 65, 196, 167, 211, 64, 223, 151, 33,
@@ -360,8 +397,8 @@ mod tests {
 							120, 131, 62, 53, 0, 54, 72, 49, 196, 234, 239, 65, 25, 159, 245, 38,
 							193, 0,
 						],
-					},
-					Cell {
+					}),
+					Cell::SingleCell(SingleCell {
 						position: Position { row: 0, col: 3 },
 						content: [
 							150, 6, 83, 12, 56, 17, 0, 225, 186, 238, 151, 181, 116, 1, 34, 240,
@@ -371,7 +408,7 @@ mod tests {
 							65, 162, 162, 190, 205, 20, 95, 67, 114, 73, 59, 170, 52, 243, 140,
 							237, 0,
 						],
-					},
+					}),
 				];
 
 				let stats = network::FetchStats::new(
@@ -418,7 +455,7 @@ mod tests {
 			.returning(move |_, _, _, _, positions| {
 				let unfetched = vec![Position { row: 0, col: 3 }];
 				let dht_fetched: Vec<Cell> = vec![
-					Cell {
+					Cell::SingleCell(SingleCell {
 						position: Position { row: 0, col: 0 },
 						content: [
 							183, 56, 112, 134, 157, 186, 15, 255, 245, 173, 188, 37, 165, 224, 226,
@@ -427,8 +464,8 @@ mod tests {
 							37, 200, 236, 4, 44, 40, 4, 3, 0, 11, 35, 249, 222, 81, 135, 1, 128, 0,
 							0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
 						],
-					},
-					Cell {
+					}),
+					Cell::SingleCell(SingleCell {
 						position: Position { row: 0, col: 2 },
 						content: [
 							153, 31, 34, 70, 221, 239, 97, 236, 3, 172, 44, 167, 114, 117, 186,
@@ -438,8 +475,8 @@ mod tests {
 							54, 133, 203, 162, 73, 252, 32, 42, 175, 24, 166, 142, 72, 226, 150,
 							163, 206, 115, 0,
 						],
-					},
-					Cell {
+					}),
+					Cell::SingleCell(SingleCell {
 						position: Position { row: 1, col: 1 },
 						content: [
 							146, 211, 61, 65, 166, 68, 252, 65, 196, 167, 211, 64, 223, 151, 33,
@@ -449,9 +486,9 @@ mod tests {
 							120, 131, 62, 53, 0, 54, 72, 49, 196, 234, 239, 65, 25, 159, 245, 38,
 							193, 0,
 						],
-					},
+					}),
 				];
-				let rpc_fetched: Vec<Cell> = vec![Cell {
+				let rpc_fetched: Vec<Cell> = vec![Cell::SingleCell(SingleCell {
 					position: Position { row: 0, col: 3 },
 					content: [
 						150, 6, 83, 12, 56, 17, 0, 225, 186, 238, 151, 181, 116, 1, 34, 240, 174,
@@ -460,7 +497,7 @@ mod tests {
 						201, 218, 107, 88, 0, 87, 199, 169, 98, 172, 4, 140, 151, 65, 162, 162,
 						190, 205, 20, 95, 67, 114, 73, 59, 170, 52, 243, 140, 237, 0,
 					],
-				}];
+				})];
 
 				let stats = network::FetchStats::new(
 					positions.len(),